+++ loguru/_datetime.py
@@ -4,9 +4,9 @@ from datetime import datetime as datetime_
 from datetime import timedelta, timezone
 from time import localtime, strftime
 
-tokens = r"H{1,2}|h{1,2}|m{1,2}|s{1,2}|S{1,6}|YYYY|YY|M{1,4}|D{1,4}|Z{1,2}|zz|A|X|x|E|Q|dddd|ddd|d"
+tokens = r"H{1,2}|h{1,2}|m{1,2}|s{1,2}|S+|YYYY|YY|M{1,4}|D{1,4}|Z{1,2}|zz|A|X|x|E|Q|dddd|ddd|d"
 
-pattern = re.compile(r"(?:{0})|\[(?:{0}|!UTC)\]".format(tokens))
+pattern = re.compile(r"(?:{0})|\[(?:{0}|!UTC|)\]".format(tokens))
 
 
 class datetime(datetime_):  # noqa: N801
@@ -22,6 +22,13 @@ class datetime(datetime_):  # noqa: N801
 
         if "%" in spec:
             return datetime_.__format__(dt, spec)
+
+        if "SSSSSSS" in spec:
+            raise ValueError(
+                "Invalid time format: the provided format string contains more than six successive "
+                "'S' characters. This may be due to an attempt to use nanosecond precision, which "
+                " not supported."
+            )
 
         year, month, day, hour, minute, second, weekday, yearday, _ = dt.timetuple()
         microsecond = dt.microsecond